+++ src/old_protocol.h
@@ -24,4 +24,4 @@ extern void old_protocol_init(int rate);
 extern void old_protocol_set_mic_sample_rate(int rate);
 
 extern void old_protocol_audio_samples(short left_audio_sample, short right_audio_sample);
-extern void old_protocol_iq_samples(int isample, int qsample, int side);
+extern void old_protocol_iq_samples(double isample, double qsample, int side);
